server-deploy/py-api/requirementsenv.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.128.5
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.3.0
pydantic==2.12.5
pydantic-core==2.41.5
starlette==0.49.3
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.39.0
opencv-python
albumentations
transformers
